Ooops -- when copying and pasting code snippets, it helps to have the right types.
git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@1117 b64f7644-9d1e-0410-96f1-a4d463321fa5

+++ libhb/deca52.c
@@ -70,7 +70,7 @@ static hb_buffer_t * Decode( hb_work_object_t * w );
  **********************************************************************/
 static sample_t dynrng_call (sample_t c, void *data)
 {        
-        double *level = (double *)data;
+        float *level = (float *)data;
         float levelToUse = (float)*level;
         if(c > 1.0 || levelToUse <= 1.0)
         {
